+++ addons/mass_mailing_sms/models/__init__.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import link_tracker
 from . import mailing_contact
 from . import mailing_list
 from . import mailing_mailing
